Extract CompactionTaskTerminateTest
