+++ java/compaction/compaction-job-execution/src/test/java/sleeper/compaction/job/execution/CompactionTaskTerminateTest.java
@@ -1,0 +1,218 @@
+/*
+ * Copyright 2022-2024 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package sleeper.compaction.job.execution;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+
+import sleeper.compaction.job.CompactionJob;
+
+import java.time.Duration;
+import java.time.Instant;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Queue;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static sleeper.configuration.properties.instance.CompactionProperty.COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS;
+import static sleeper.configuration.properties.instance.CompactionProperty.COMPACTION_TASK_MAX_CONSECUTIVE_FAILURES;
+import static sleeper.configuration.properties.instance.CompactionProperty.COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS;
+
+public class CompactionTaskTerminateTest extends CompactionTaskTestBase {
+
+    @Nested
+    @DisplayName("Stop if idle for a specified period")
+    class StopAfterMaxIdleTime {
+
+        @Test
+        void shouldTerminateIfNoJobsArePresentAfterRunningForIdleTime() throws Exception {
+            // Given
+            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
+            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
+            Queue<Instant> times = new LinkedList<>(List.of(
+                    Instant.parse("2024-02-22T13:50:00Z"), // Start
+                    Instant.parse("2024-02-22T13:50:03Z"))); // Finish
+
+            // When
+            runTask(processNoJobs(), times::poll);
+
+            // Then
+            assertThat(times).isEmpty();
+            assertThat(sleeps).isEmpty();
+        }
+
+        @Test
+        void shouldTerminateIfNoJobsArePresentAfterRunningForIdleTimeWithTwoQueuePolls() throws Exception {
+            // Given
+            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
+            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
+            Queue<Instant> times = new LinkedList<>(List.of(
+                    Instant.parse("2024-02-22T13:50:00Z"), // Start
+                    Instant.parse("2024-02-22T13:50:02Z"), // First idle time check
+                    Instant.parse("2024-02-22T13:50:04Z"))); // Second idle time check + finish
+
+            // When
+            runTask(processNoJobs(), times::poll);
+
+            // Then
+            assertThat(times).isEmpty();
+            assertThat(sleeps).containsExactly(Duration.ofSeconds(2));
+        }
+
+        @Test
+        void shouldTerminateAfterRunningJobAndWaitingForIdleTime() throws Exception {
+            // Given
+            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
+            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
+            Queue<Instant> times = new LinkedList<>(List.of(
+                    Instant.parse("2024-02-22T13:50:00Z"), // Start
+                    Instant.parse("2024-02-22T13:50:01Z"), // Job started
+                    Instant.parse("2024-02-22T13:50:02Z"), // Job completed
+                    Instant.parse("2024-02-22T13:50:05Z"))); // Idle time check with empty queue and finish
+            CompactionJob job = createJobOnQueue("job1");
+
+            // When
+            runTask(jobsSucceed(1), times::poll);
+
+            // Then
+            assertThat(times).isEmpty();
+            assertThat(successfulJobs).containsExactly(job);
+            assertThat(failedJobs).isEmpty();
+            assertThat(jobsOnQueue).isEmpty();
+            assertThat(sleeps).isEmpty();
+        }
+
+        @Test
+        void shouldTerminateWhenMaxIdleTimeNotMetOnFirstCheckThenIdleAfterProcessingJob() throws Exception {
+            // Given
+            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
+            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
+            Queue<Instant> times = new LinkedList<>(List.of(
+                    Instant.parse("2024-02-22T13:50:00Z"), // Start
+                    Instant.parse("2024-02-22T13:50:01Z"), // First check
+                    Instant.parse("2024-02-22T13:50:02Z"), // Job started
+                    Instant.parse("2024-02-22T13:50:02Z"), // Job completed
+                    Instant.parse("2024-02-22T13:50:06Z"))); // Second check + finish
+            CompactionJob job = createJob("job1");
+
+            // When
+            runTask(
+                    pollQueue(
+                            receiveNoJobAnd(() -> send(job)),
+                            receiveJob(),
+                            receiveNoJob()),
+                    processJobs(jobSucceeds()),
+                    times::poll);
+
+            // Then
+            assertThat(times).isEmpty();
+            assertThat(successfulJobs).containsExactly(job);
+            assertThat(failedJobs).isEmpty();
+            assertThat(jobsOnQueue).isEmpty();
+            assertThat(sleeps).containsExactly(Duration.ofSeconds(2));
+        }
+
+        @Test
+        void shouldTerminateWhenMaxIdleTimeNotMetOnFirstCheckThenNotMetAfterProcessingJob() throws Exception {
+            // Given
+            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
+            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
+            Queue<Instant> times = new LinkedList<>(List.of(
+                    Instant.parse("2024-02-22T13:50:00Z"), // Start
+                    Instant.parse("2024-02-22T13:50:01Z"), // First check
+                    Instant.parse("2024-02-22T13:50:02Z"), // Job started
+                    Instant.parse("2024-02-22T13:50:03Z"), // Job completed
+                    Instant.parse("2024-02-22T13:50:04Z"), // Second check
+                    Instant.parse("2024-02-22T13:50:06Z"))); // Third check + finish
+            CompactionJob job = createJob("job1");
+
+            // When
+            runTask(
+                    pollQueue(
+                            receiveNoJobAnd(() -> send(job)),
+                            receiveJob(),
+                            receiveNoJob(),
+                            receiveNoJob()),
+                    processJobs(jobSucceeds()),
+                    times::poll);
+
+            // Then
+            assertThat(times).isEmpty();
+            assertThat(successfulJobs).containsExactly(job);
+            assertThat(failedJobs).isEmpty();
+            assertThat(jobsOnQueue).isEmpty();
+            assertThat(sleeps).containsExactly(Duration.ofSeconds(2), Duration.ofSeconds(2));
+        }
+
+        @Test
+        void shouldNotDelayRetryIfSetToZero() throws Exception {
+            // Given
+            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
+            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 0);
+            Queue<Instant> times = new LinkedList<>(List.of(
+                    Instant.parse("2024-02-22T13:50:00Z"), // Start
+                    Instant.parse("2024-02-22T13:50:02Z"), // First idle time check
+                    Instant.parse("2024-02-22T13:50:04Z"))); // Second idle time check + finish
+
+            // When
+            runTask(processNoJobs(), times::poll);
+
+            // Then
+            assertThat(times).isEmpty();
+            assertThat(sleeps).isEmpty();
+        }
+    }
+
+    @Nested
+    @DisplayName("Stop if failed too many times consecutively")
+    class StopAfterConsecutiveFailures {
+        @Test
+        void shouldStopEarlyIfMaxConsecutiveFailuresMet() throws Exception {
+            // Given
+            instanceProperties.setNumber(COMPACTION_TASK_MAX_CONSECUTIVE_FAILURES, 2);
+            CompactionJob job1 = createJobOnQueue("job1");
+            CompactionJob job2 = createJobOnQueue("job2");
+            CompactionJob job3 = createJobOnQueue("job3");
+
+            // When
+            runTask(processJobs(jobFails(), jobFails(), jobSucceeds()));
+
+            // Then
+            assertThat(successfulJobs).isEmpty();
+            assertThat(failedJobs).containsExactly(job1, job2);
+            assertThat(jobsOnQueue).containsExactly(job3);
+        }
+
+        @Test
+        void shouldResetConsecutiveFailureCountIfJobProcessedSuccessfully() throws Exception {
+            // Given
+            instanceProperties.setNumber(COMPACTION_TASK_MAX_CONSECUTIVE_FAILURES, 2);
+            CompactionJob job1 = createJobOnQueue("job1");
+            CompactionJob job2 = createJobOnQueue("job2");
+            CompactionJob job3 = createJobOnQueue("job3");
+            CompactionJob job4 = createJobOnQueue("job4");
+
+            // When
+            runTask(processJobs(jobFails(), jobSucceeds(), jobFails(), jobSucceeds()));
+
+            // Then
+            assertThat(successfulJobs).containsExactly(job2, job4);
+            assertThat(failedJobs).containsExactly(job1, job3);
+            assertThat(jobsOnQueue).isEmpty();
+        }
+    }
+}
