+++ java/compaction/compaction-job-execution/src/test/java/sleeper/compaction/job/execution/CompactionTaskTest.java
@@ -26,7 +26,6 @@ import sleeper.compaction.task.CompactionTaskStatus;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
 
-import java.time.Duration;
 import java.time.Instant;
 import java.util.LinkedList;
 import java.util.List;
@@ -37,9 +36,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.compaction.job.CompactionJobStatusTestData.finishedCompactionRun;
 import static sleeper.compaction.job.CompactionJobStatusTestData.jobCreated;
 import static sleeper.compaction.job.CompactionJobStatusTestData.startedCompactionRun;
-import static sleeper.configuration.properties.instance.CompactionProperty.COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS;
-import static sleeper.configuration.properties.instance.CompactionProperty.COMPACTION_TASK_MAX_CONSECUTIVE_FAILURES;
-import static sleeper.configuration.properties.instance.CompactionProperty.COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS;
 
 public class CompactionTaskTest extends CompactionTaskTestBase {
 
@@ -87,188 +83,6 @@ public class CompactionTaskTest extends CompactionTaskTestBase {
             // Then
             assertThat(successfulJobs).containsExactly(job1);
             assertThat(failedJobs).containsExactly(job2);
-            assertThat(jobsOnQueue).isEmpty();
-        }
-    }
-
-    @Nested
-    @DisplayName("Stop if idle for a specified period")
-    class StopAfterMaxIdleTime {
-
-        @Test
-        void shouldTerminateIfNoJobsArePresentAfterRunningForIdleTime() throws Exception {
-            // Given
-            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
-            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:03Z"))); // Finish
-
-            // When
-            runTask(processNoJobs(), times::poll);
-
-            // Then
-            assertThat(times).isEmpty();
-            assertThat(sleeps).isEmpty();
-        }
-
-        @Test
-        void shouldTerminateIfNoJobsArePresentAfterRunningForIdleTimeWithTwoQueuePolls() throws Exception {
-            // Given
-            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
-            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:02Z"), // First idle time check
-                    Instant.parse("2024-02-22T13:50:04Z"))); // Second idle time check + finish
-
-            // When
-            runTask(processNoJobs(), times::poll);
-
-            // Then
-            assertThat(times).isEmpty();
-            assertThat(sleeps).containsExactly(Duration.ofSeconds(2));
-        }
-
-        @Test
-        void shouldTerminateAfterRunningJobAndWaitingForIdleTime() throws Exception {
-            // Given
-            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
-            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:01Z"), // Job started
-                    Instant.parse("2024-02-22T13:50:02Z"), // Job completed
-                    Instant.parse("2024-02-22T13:50:05Z"))); // Idle time check with empty queue and finish
-            CompactionJob job = createJobOnQueue("job1");
-
-            // When
-            runTask(jobsSucceed(1), times::poll);
-
-            // Then
-            assertThat(times).isEmpty();
-            assertThat(successfulJobs).containsExactly(job);
-            assertThat(failedJobs).isEmpty();
-            assertThat(jobsOnQueue).isEmpty();
-            assertThat(sleeps).isEmpty();
-        }
-
-        @Test
-        void shouldTerminateWhenMaxIdleTimeNotMetOnFirstCheckThenIdleAfterProcessingJob() throws Exception {
-            // Given
-            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
-            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:01Z"), // First check
-                    Instant.parse("2024-02-22T13:50:02Z"), // Job started
-                    Instant.parse("2024-02-22T13:50:02Z"), // Job completed
-                    Instant.parse("2024-02-22T13:50:06Z"))); // Second check + finish
-            CompactionJob job = createJob("job1");
-
-            // When
-            runTask(
-                    pollQueue(
-                            receiveNoJobAnd(() -> send(job)),
-                            receiveJob(),
-                            receiveNoJob()),
-                    processJobs(jobSucceeds()),
-                    times::poll);
-
-            // Then
-            assertThat(times).isEmpty();
-            assertThat(successfulJobs).containsExactly(job);
-            assertThat(failedJobs).isEmpty();
-            assertThat(jobsOnQueue).isEmpty();
-            assertThat(sleeps).containsExactly(Duration.ofSeconds(2));
-        }
-
-        @Test
-        void shouldTerminateWhenMaxIdleTimeNotMetOnFirstCheckThenNotMetAfterProcessingJob() throws Exception {
-            // Given
-            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
-            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 2);
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:01Z"), // First check
-                    Instant.parse("2024-02-22T13:50:02Z"), // Job started
-                    Instant.parse("2024-02-22T13:50:03Z"), // Job completed
-                    Instant.parse("2024-02-22T13:50:04Z"), // Second check
-                    Instant.parse("2024-02-22T13:50:06Z"))); // Third check + finish
-            CompactionJob job = createJob("job1");
-
-            // When
-            runTask(
-                    pollQueue(
-                            receiveNoJobAnd(() -> send(job)),
-                            receiveJob(),
-                            receiveNoJob(),
-                            receiveNoJob()),
-                    processJobs(jobSucceeds()),
-                    times::poll);
-
-            // Then
-            assertThat(times).isEmpty();
-            assertThat(successfulJobs).containsExactly(job);
-            assertThat(failedJobs).isEmpty();
-            assertThat(jobsOnQueue).isEmpty();
-            assertThat(sleeps).containsExactly(Duration.ofSeconds(2), Duration.ofSeconds(2));
-        }
-
-        @Test
-        void shouldNotDelayRetryIfSetToZero() throws Exception {
-            // Given
-            instanceProperties.setNumber(COMPACTION_TASK_MAX_IDLE_TIME_IN_SECONDS, 3);
-            instanceProperties.setNumber(COMPACTION_TASK_DELAY_BEFORE_RETRY_IN_SECONDS, 0);
-            Queue<Instant> times = new LinkedList<>(List.of(
-                    Instant.parse("2024-02-22T13:50:00Z"), // Start
-                    Instant.parse("2024-02-22T13:50:02Z"), // First idle time check
-                    Instant.parse("2024-02-22T13:50:04Z"))); // Second idle time check + finish
-
-            // When
-            runTask(processNoJobs(), times::poll);
-
-            // Then
-            assertThat(times).isEmpty();
-            assertThat(sleeps).isEmpty();
-        }
-    }
-
-    @Nested
-    @DisplayName("Stop if failed too many times consecutively")
-    class StopAfterConsecutiveFailures {
-        @Test
-        void shouldStopEarlyIfMaxConsecutiveFailuresMet() throws Exception {
-            // Given
-            instanceProperties.setNumber(COMPACTION_TASK_MAX_CONSECUTIVE_FAILURES, 2);
-            CompactionJob job1 = createJobOnQueue("job1");
-            CompactionJob job2 = createJobOnQueue("job2");
-            CompactionJob job3 = createJobOnQueue("job3");
-
-            // When
-            runTask(processJobs(jobFails(), jobFails(), jobSucceeds()));
-
-            // Then
-            assertThat(successfulJobs).isEmpty();
-            assertThat(failedJobs).containsExactly(job1, job2);
-            assertThat(jobsOnQueue).containsExactly(job3);
-        }
-
-        @Test
-        void shouldResetConsecutiveFailureCountIfJobProcessedSuccessfully() throws Exception {
-            // Given
-            instanceProperties.setNumber(COMPACTION_TASK_MAX_CONSECUTIVE_FAILURES, 2);
-            CompactionJob job1 = createJobOnQueue("job1");
-            CompactionJob job2 = createJobOnQueue("job2");
-            CompactionJob job3 = createJobOnQueue("job3");
-            CompactionJob job4 = createJobOnQueue("job4");
-
-            // When
-            runTask(processJobs(jobFails(), jobSucceeds(), jobFails(), jobSucceeds()));
-
-            // Then
-            assertThat(successfulJobs).containsExactly(job2, job4);
-            assertThat(failedJobs).containsExactly(job1, job3);
             assertThat(jobsOnQueue).isEmpty();
         }
     }
